Poison
I added poison to the shop and its effect as well.

+++ DrakeShot-main/Drake Shot/Assets/scripts/ShopItemScript.cs
@@ -5,7 +5,7 @@ public class ShopItemScript : MonoBehaviour {
     
     [SerializeField] private TextMeshPro addText;
     [SerializeField] private int price;
-    [SerializeField] private int effectValue;
+    [SerializeField] private float effectValue;
     [SerializeField] private bool isRepurchasable;
 
     private BuyScript _myBuyScript;
